site

+++ Controller/SCPController.php
@@ -2,7 +2,6 @@
 
 require_once("Models/scpModel.php");
 if($uri === "/scp"){
-
     $safe = selectAllSCPsOnClass($pdo, "Safe");
     $euclide = selectAllSCPsOnClass($pdo, "Euclide");
     $keter = selectAllSCPsOnClass($pdo, "Keter");
@@ -13,25 +12,38 @@ if($uri === "/scp"){
     require_once("Views/base.php");
 
 } else if($uri === "/scpnew"){
-    if(isset($_POST['btnEnvoi'])){
+    if(isset($_SESSION["user"])){
+        if($_SESSION["user"]->userPermission === "admin"){
+            if(isset($_POST['btnEnvoi'])){
+                
+                $messageError = verifEmptyData();
+                if(createSCP($pdo)){
+                    
+                    header('location:/scp');
+                } else{
         
-        $messageError = verifEmptyData();
-        if(createSCP($pdo)){
+                }
+                
+            }else{
+                $sites = selectAllSites($pdo);
+                $title = "Ajouter un SCP";
+                $template = "Views/SCP/SCPaddOrEdit.php"; 
+                require_once("Views/base.php");
+                
+            }
             
-            header('location:/scp');
         } else{
-
+            header("location:/scp");
         }
-        
-    }else{
-        $sites = selectAllSites($pdo);
-        $title = "Ajouter un SCP";
-        $template = "Views/SCP/SCPaddOrEdit.php"; 
-        require_once("Views/base.php");
-        
+    } else{
+        header("location:/scp");
     }
 } else if(isset($_GET["SCPID"]) && $uri === "/scpDelete?SCPID=" . $_GET["SCPID"]){
-    deleteSCPFromID($pdo);
+    if(isset($_SESSION["user"])){
+        if($_SESSION["user"]->userPermission === "admin"){
+            deleteSCPFromID($pdo);
+        }
+    }
     header('location:/scp');
